# report: optimize newTrimmedGraph allocs